Deterministic order of TmplObjectArray function
Template function objectArray did not order values deterministic.
This changes ensures this by ordering after all values are parsed.

The test assertion is changed to `assert.Equal` that matches the values
directly instead of `assert.ElementsMatch` that does not take order into
account.
With this change the "nil input" test case did not match expectatations.
This is due to `assert.ElementsMatch` accepting a nil value to match an
empty slice.
This means the output of the function have not changed, but our
assertion has gotten more strict.

Related to #29