Deterministic order of TmplArray function
Template function array did not order values deterministic.
This change ensures this for string arrays.
We have no trivial way of ordering unknown types so this is omitted.

I removed some code handling objects in the input.
The function is not intended to handle these cases.
TmplObjectArray should be used instead.

Related to #29