Handle templates not co located in projects
If a shuttle template command is run on a template not inside the
project the template name does not resolve to the rigth template.

This change ensures to execute the named template based on the
path's base name.

I refactored the output handling as well to avoid having two
tmpl.Execute calls.