Fix stdout and stderr handling in sub commands
Currently the shell executor and gitCmd helper writes exessive newlines during
execution after the upgrade of go-cmd/cmd where handling of the Stderr and
Stdout channels was changed.

The update was introduced in dd2f8b36d93a48a7268b39d09df137a09651ddde to fix
missing output of git-plan commands but was mistakenly not ported to the other
uses of the cmd package.

This change set updates usage of package cmd when configured with streaming
output.