Add run flag for skipping argument validation
Currently it is not possible to skip argument validation introduced in
c68e9b7230c7ec0cd25f761ccf27f61f7ca6ca98. This imposes an inconvinience when
using script overwrites to disable shared functionality.

Example:

A project uses plan 'foo' with a script 'bar' that requires argument 'baz' and
it is used in a shared pipeline, eg. Jenkinsfile, with the argument set
correctly.

If the project for some reason wants to opt-out of the script behaviour it can
add a local script of the same name 'bar' and define it as 'echo skipping...'.

With argument validation enabled in the pipeline, the project is no longer able
to build and maintainers of the plan can no longer add new required arguments to
the shared script as that would break any one opting out.

As a solution to this, the 'validate' flag is added to the 'run' command. It is
intended to be used in shared invocations of scripts, eg. pipelines, and thus
gives the usual flexibility while still giving local development validation by
default.

The reasoning for this compromise is the validation has the most effect on local
development where developers might not know what arguments should be supplied or
or maybe they mistype their names. Developers working on pipelines take the hit,
but as that is a more slow moving target we accept that.