Remove os.Exit calls in pkg layer (#61)
Currently a UI instance is passed around to all parts of shuttle allowing layers
to report messages to the user. It also provides methods ExitWithError,
ExitWithErrorCode and CheckIfError which are fatal messages stopping shuttle
completely by calling os.Exit. This makes shuttle in it self hard to test and
provides unclear exit paths of the code.

This change contains multiple changes to fix above issues.

The methods from package ui are removed and instead a checkError function is
added to package cmd which is responsible for selecting an exit code based on
the error. This greatly improvoes maintainability of the code and makes it
easier to reason about any possible exit paths from code in packages of pkg/*.

Some dead code is removed from package pkg/config. An unused CheckIfError
function was laying around here and as this change moves functionality around
related to the nature of this one I figured it was worth removing now.

During the change I found a couple of bugs in the error reporting that is fixed
as well.

When using command 'run' with invalid arguments exit code 1 indicating a shuttle
failure is reported instead of 2 indicating bad user input.

If a plan cannot be fetched due to an unknown protocol a panic is thrown. It now
returns an ExitCode error indicating bad user input.

Errors around the .shuttle directory are now also detected and propagated
instead of silently ignored.