Listen to SIGINT and close scripts
Currently if a shuttle script is cancelled by sending a SIGINT signal the
process running in a script is not signalled. This leads to orphaned processes
and is especially an issue when launching web servers listening on a port that
can only by used once. Here the second run of the script will fail as the port
is already used.

This change introduces a context.Context down through the call chain and the
shell executor will listen for cancellations on this and signal the processes
accordingly. The context is cancelled when receiving SIGINT.

Module github.com/go-cmd/cmd is updated to the latest version which contains a
more stable shutdown API.