Add documentation command
This change set contains a new documentation command.

    $ shuttle docs --project examples/moon-base
    Documentation available at: https://github.com/lunarway/shuttle

It will open up documentation for the used shuttle plan. If a plan specifices a
top level "documentation" field, that is used. If not available the plan
reference is parsed and a URL is inferred.

The command will open the found URL respecting the BROWSER environment variable
and falling back to OS specific ways of "opening" URLs.