Handle large stdout and stderr lines from scripts (#70)
Currently if a script outputs a single line to stdout or stderr over 16k
characters the script is stopped with a pipe error. When used in combination
with shell scripts this makes for a very confusing error state where output is
just limited or the process either hangs or fails without explanation.

This change replaces the shell executer based on github.com/go-cmd/cmd with one
using the standard libraries os/exec package. Further more the line buffer limit
is increased to 512k characters.

go-cmd does not support changing the maximum line length so the library cannot
support our use case.

This bug was found when a Go service was testing an HTML template. The test
would output the rendered HTML template and a complex SVG exceeded the current
16k limit. This resulted in the test being marked as successfull by the Go
toolchain but as the stdout pipe was broken shell scripts would not work as
intended.