Support shuttle commands from sub directories
Currently when shuttle is invoked it needs to be in the root of a project. This
limits its uses in scripts like go:generate where files in sub directories would
rely on the root project path.

This change adds a recursive lookup of the project path by walking backwards in
the file tree to find a shuttle.yaml file. This is a well known pattern from
other CLIs like Git where a CLI command from a sub directory acts as if it was
invoked from the project root.