Fail template rendering on template func errors
Currently when using template functions toYaml, fromYaml and getFileContent
fails silently or in odd ways where the rendered output contains error
information. Go's template engine supports template functions that returns
errors and thus will propagate errors correctly. This can help identify falty
paths and contents in a clear way instead of silently doing nothing.

The change is motivated by fighting with understanding why getFileContent does
nothing in some cases. As the error is hidden there are no hints as to what is
going on.

This change adds error return values to the functions along with adds some tests
of this behaviour. The functions are considered safe to change as it is unusual
to use them without expeting failures on bad input, eg. read a non-existing
file. Function getFiles is not changed to return an error as it is expected that
a non-existing or empty directory will return no files. A test is added to lock
this behaviour.