Remove type assertions in TmplObjectArray
Currently a type switch on a value in TmplObjectArray doesn't store the typed
value in the switch. This means that all cases needs to cast to the same time.
This is both error prone, eg. casting to the wrong type, and also required more
typing.

This change stored the type of the switched value which removes the error risk
and additional type casting.

This optimization was reported by go-staticcheck rule S1034.