Migrate to Github Actions (#86)
This change migrates our CI solution to GitHub actions from Travis CI.

The single flow Travis pipeline is split into a build and release actions
workflow. Caching of Go modules is added to the build flow and changelogs are
now managed by release-drafter instead of GoReleaser.

The shunit2 tests does not work in GitHub actions so we skip them for now and
will make an effort on migrating them to Go.