Remove global cobra command state (#88)
Currently all the cobra commands are stored with a global state making
dependencies and order of execution impossible to understand.

This change adds constructors to all commands with arguments for dependencies
unwrapping the global state.

The change is motivated by spurious test failures depending on what order they
are run in. Further more it is a general quiality enhancement of the code base
leading the way for more general refactors.