Inject output streams and avoid copying ui.UI (#91)
Currently the ui.UI type copies it self when running the build methods which
makes it hard to reason about configuraiton of its values in different parts of
the code base.

This change makes the builder pattern mutable while updating all usage to be
pointer based. This makes sure changes to the UI propagates correctly. Further
more the output streams are now injected at construction time to force setting
it and the cmd.Execute function is updated to take these as inputs.

All in all this makes sure we actually output everything to the configured
streams.