Remove os.Exit from commands (#92)
Currently the checkError function is used through out all the command's handlers
which makes them exit without cobra handling it. It also makes the commands very
hard to test.

As the root commands Execute already inspects the errors returned from
individual command handlers and also detects any ExitCode errors this change is
a noop. All handlers now return any error they encounter and let it be up to the
client of Execute to decide what to do with it.

Some tests of the behaviour is added to the command test suite with a new
dedicated test plan instead of the arbitrary projects and plans inside the
"examples" directory.