Fix shell output race condition (#107)
This change partially reverts 48c5180f2a33e5766199ea54ca29cd5c587dd7f6 (#70)
which fixed an issue where large outputs from shell scripts would break output
pipes (#90).

The change replaced go-cmd with a native implementation which on the surface
fixed the large output bug but introduced another subtle race condition in the
output of scripts.

The motivation in the original commit for replacing go-cmd was the missing
ability to configure the output buffers size. That feature is now available in
go-cmd why this change set reverts to use the proven go-cmd implementation. This
removes a lot of complexity from shuttle while also fixing the race condition
identified in #90.