Run all tests in test-go (#108)
Currently only Go tests in package pkg is tested in the test-go script. This
means the command accept tests in package cmd is not run.

This change fixes that.

With the change it became clear that tests relied on other tests so additional
cleanup is added to the test cases making sure that any .shuttle directories are
removed before and after each test case. Some tests are also updated to use
testdata instead of examples to make the consistent and make the cleanup work.