Remove global state from executors package (#112)
Currently package executors contains a global slice of actionExector structs
that each executor registeres to with init functions. This makes them hard to
test and tightly coupled along with making it hard to understand the code.

This change introduces a Registry struct that is newed up in the run command
along with selecting what executors should be used. The init functions are
replaced with exported matchers.

A small finding. The docker executor is actually not working but is still
available. This one is not registered in the new registry as the in-complete
implementation will be removed in another change.