Fix some minor formatting on config command
There are some odd whitespaces in the config command output and missing line
breaks if no plan is used.

This change expands the test cases to cover these scenarios which located a big
in the gitCmd function in package git where an os/exec error would go unnoticed.

This change fixes that bug along with the formatting. A change to how the tests
are run was needed to make assertions environment agnostic as we cannot clear
the environment variabls up front as that breakes stuff like PATH used for plans
with git.